requirements.txt
Package         Version
--------------- -------
asyncpg         0.22.0
attrs           20.3.0
click           7.1.2
colorama        0.4.4
cycler          0.10.0
databases       0.4.1
fastapi         0.63.0
fastapi-restful 0.2.5
h11             0.12.0
httptools       0.1.1
iniconfig       1.1.1
kiwisolver      1.3.1
matplotlib      3.3.4
numpy           1.20.1
packaging       20.9
pandoc          1.0.2
Pillow          8.1.2
pip             20.2.3
pluggy          0.13.1
ply             3.11
py              1.10.0
pydantic        1.7.3
pyparsing       2.4.7
pytest          6.2.2
python-dateutil 2.8.1
python-dotenv   0.15.0
PyYAML          5.4.1
setuptools      49.2.1
shellingham     1.4.0
six             1.15.0
SQLAlchemy      1.3.23
starlette       0.13.6
toml            0.10.2
typer           0.3.2
uvicorn         0.13.4
uvloop          0.15.2
watchgod        0.7
websockets      8.1
